+++ src/control_plane/state/credential_record_options.rs
@@ -4,7 +4,7 @@ use super::CredentialLifecycle;
 #[derive(Clone, Eq, PartialEq)]
 pub(crate) struct CredentialRecordOptions {
     pub(crate) credential_id: String,
-    pub(crate) project_id: String,
+    pub(crate) project_id: Option<String>,
     pub(crate) service_id: String,
     pub(crate) username: String,
     pub(crate) secret: String,